Remove old spark-local
